chore(lib::config): move all config code to a v1 module
